Move event detail files into own folder
