+++ src/screens/eventDetail/EventDetailsItemContainer.tsx
