+++ src/screens/eventDetail/EventDetailsViewModel.tsx
