+++ src/screens/eventDetail/EventDetailsViewModelMapper.ts
