+++ src/screens/events/EventNavigator.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react'
 import { createStackNavigator } from '@react-navigation/stack'
 import { EventParamList, Screen } from '../../navigation'
-import EventDetailsScreen from './EventDetailsScreen'
+import EventDetailsScreen from '../eventDetail/EventDetailsScreen'
 import { headerBlank } from '../../styles/navigationAppearance'
 import EventsScreen from './EventsScreen'
